Build fixes for g++ 4.4.7 on CentOS 6.
M    configure.ac
M    conf/libtool.m4
M    conf/ltversion.m4
M    conf/depcomp
M    conf/missing
M    conf/ltoptions.m4
M    conf/config.guess
M    conf/ltmain.sh
M    conf/config.sub
M    conf/install-sh
M    src/rankeddataset.h
M    src/expr.cc
M    src/tuple.cc
M    src/cellarray.cc
M    src/grid.cc

+++ trunk/clib/src/rankeddataset.h
@@ -42,7 +42,7 @@ class RankedDataset : public Object {
 
    void Clear() { ranks.clear(); }
 
-   void CoerceScheme(Dim_t k, Scheme *sch, size_t sz=0) { assert(k<ranks.size()); return ranks[k].CoerceScheme(*sch, sz);};
+   void CoerceScheme(Dim_t k, Scheme *sch, size_t sz=0) { assert((unsigned)k<ranks.size()); return ranks[k].CoerceScheme(*sch, sz);};
   
    void AddAttribute(Dim_t k, Array *data) { 
      if (k>=(Dim_t)ranks.size()) {
